major: refactor and organization
1. Clear css files since we will be using bootstrap, we might use
index.css for custom mobile theme later.
2. Organized files, compoenets directory will be for parts of a single
page, and pages folder will contain full pages. Started with a single
component (navbar)
3. App.jsx its the main application, holding the Routes and the Navbar.

Next will be working on testing login/signup and overall backend
integration.
